requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Environment
python-dotenv==1.0.1
pydantic-settings==2.6.0

# Vector store
qdrant-client==1.12.0

# Embeddings
openai==2.17.0
cohere==5.13.0
google-generativeai==0.8.3
fastembed==0.7.4

# Document processing
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3

# Utilities
numpy==1.26.4
